server/requirements.txt
psycopg2-binary
lxml
pandas
grpcio
grpcio-tools